backend/requirements.txt
# Core Django
Django>=5.0,<5.3
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.0

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# API Documentation
drf-spectacular>=0.27.0

# CORS
django-cors-headers>=4.0.0

# Additional dependencies (auto-installed with above packages)
# asgiref, sqlparse, PyJWT, etc.
